+++ src/modules/appointments/http/controllers/Appointments.controller.ts
@@ -1,0 +1,21 @@
+import { Request, Response } from 'express';
+import { parseISO } from 'date-fns';
+import { container } from 'tsyringe';
+import CreateAppointmentService from '@modules/appointments/services/CreateAppointmentService';
+
+export default class AppointmentsController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { date, providerId } = request.body;
+
+    const parseDate = parseISO(date);
+
+    const createAppointment = container.resolve(CreateAppointmentService);
+
+    const appointment = await createAppointment.execute({
+      date: parseDate,
+      providerId,
+    });
+
+    return response.json(appointment);
+  }
+}
