+++ src/modules/users/http/controllers/UserAvatar.controller.ts
@@ -1,0 +1,18 @@
+import { Request, Response } from 'express';
+import { container } from 'tsyringe';
+import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
+
+export default class UserAvatarController {
+  public async update(request: Request, response: Response): Promise<Response> {
+    const updateUserAvatarService = container.resolve(UpdateUserAvatarService);
+
+    const user = await updateUserAvatarService.execute({
+      userId: request.user.id,
+      avatarFileName: request.file.filename,
+    });
+
+    delete user.password;
+
+    return response.json(user);
+  }
+}
