+++ src/modules/users/http/controllers/Users.controller.ts
@@ -1,0 +1,16 @@
+import { Request, Response } from 'express';
+import { container } from 'tsyringe';
+import CreateUserService from '@modules/users/services/CreateUserService';
+
+export default class UsersController {
+  public async create(request: Request, response: Response): Promise<Response> {
+    const { name, email, password } = request.body;
+    const userService = container.resolve(CreateUserService);
+
+    const user = await userService.execute({ name, email, password });
+
+    delete user.password;
+
+    return response.json(user);
+  }
+}
