+++ src/modules/users/http/middlewares/ensureAuthenticate.middleware.ts
